+++ PoC/kafka-pull/kafka-broker-extension/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-com.zf.dos.edc.extensions.kafka.KafkaBrokerExtension
+org.eclipse.tractusx.edc.extensions.kafka.KafkaBrokerExtension
