+++ PoC/kafka-pull/validator-data-address-kafka/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-com.zf.dos.edc.validator.dataaddress.kafka.KafkaBrokerDataAddressValidatorExtension
+org.eclipse.tractusx.edc.validator.dataaddress.kafka.KafkaBrokerDataAddressValidatorExtension
